tests/wpt/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.pattern.animated.gif.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.pattern.animated.gif</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.pattern.animated.gif</h1>
<p class="desc">createPattern() of an animated GIF draws the first frame</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("createPattern() of an animated GIF draws the first frame");
_addTest(function(canvas, ctx) {

deferTest();
step_timeout(function () {
    var pattern = ctx.createPattern(document.getElementById('anim-gr.gif'), 'repeat');
    ctx.fillStyle = pattern;
    ctx.fillRect(0, 0, 50, 50);
    step_timeout(t.step_func_done(function () {
        ctx.fillRect(50, 0, 50, 50);
        _assertPixelApprox(canvas, 25,25, 0,255,0,255, "25,25", "0,255,0,255", 2);
        _assertPixelApprox(canvas, 75,25, 0,255,0,255, "75,25", "0,255,0,255", 2);
    }), 250);
}, 250);


});
</script>
<img src="/images/anim-gr.gif" id="anim-gr.gif" class="resource">
